refactor: LottoWinningNumbers와 LottoBonusNumber 분리
